refactor(billing): move the plan-credit derivation into service-contracts
A standalone package was more plumbing than code for a few pure
functions. service-contracts already holds cross-service math both the
daemon and the web import (stripe-currency), and every consumer already
depends on it, so the derivation lives there as the plan-credit
subpath. This drops packages/billing and its Dockerfile copy, knip
ignore, bundled-dependency entry, manifest changes, and lockfile diff.
The web CI path filters now include service-contracts, which the web
imports but the filters had omitted.
