+++ .github/workflows/pr-web.yaml
@@ -20,7 +20,7 @@ on:
       - 'packages/environments/**'
       - 'packages/avatar-catalog/**'
       - 'packages/avatar-manifest/**'
-      - 'packages/billing/**'
+      - 'packages/service-contracts/**'
       # The web LLM catalog parity test reads this file, so daemon-side catalog
       # regenerations must run the web tests on the PR that introduces drift.
       - 'meta/llm-provider-catalog.json'
