--- packages/billing/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@vellumai/billing",
-  "version": "0.0.1",
-  "private": true,
-  "license": "MIT",
-  "type": "module",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "typecheck": "bunx tsc --noEmit",
-    "test": "bun test src/"
-  },
-  "devDependencies": {
-    "@types/bun": "1.3.11",
-    "typescript": "5.9.3"
-  }
-}
